requirements.txt
asgiref==3.5.2
bcrypt==4.0.1
dj-database-url==1.2.0
Django==4.1.4
django-cors-headers==3.13.0
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
pytz==2022.7
sqlparse==0.4.3
whitenoise==6.3.0
